pyproject.toml
[project]
name = "deterministic-ai-agent"
version = "0.1.0"
description = "SentenceTransformer-based deterministic agent skeleton"
dependencies = [
    "numpy",
    "onnxruntime",
    "tokenizers",
    "pydantic",
    "pyyaml",
]

[project.optional-dependencies]
train = [
    "torch",
    "sentence-transformers",
    "scikit-learn",
    "pandas",
    "tqdm",
    "optimum[exporters]",
]
dev = [
    "pytest",
    "pytest-ruff",
    "pytest-mypy",
    "ruff",
    "mypy",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
ignore_missing_imports = true
strict = false

[tool.pytest.ini_options]
addopts = "--ruff --ruff-format --mypy"
testpaths = ["deterministic_ai_agent", "tests"]
